refactor(modelfit): make `tqdm` an optional dependency
The `tqdm` package which provides the progress bar when `progress=True` is now an optional dependency. If not installed, passing `progress=True` to `modelfit` will now result in an error.

BREAKING CHANGE: Using `modelfit` with `progress=True` now requires the package to be installed with the `progress` optional dependency group, like `pip install xarray-lmfit[progress]`.